+++ src/plugins/discover/public/application/apps/main/utils/fetch_all.ts
@@ -12,18 +12,18 @@ import {
   sendLoadingMsg,
   sendPartialMsg,
   sendResetMsg,
-} from './use_saved_search_messages';
-import { updateSearchSource } from '../utils/update_search_source';
+} from '../services/use_saved_search_messages';
+import { updateSearchSource } from './update_search_source';
 import { SortOrder } from '../../../../saved_searches/types';
 import { fetchDocuments } from './fetch_documents';
 import { fetchTotalHits } from './fetch_total_hits';
 import { fetchChart } from './fetch_chart';
 import { SearchSource } from '../../../../../../data/common';
 import { Adapters } from '../../../../../../inspector';
-import { AppState } from './discover_state';
+import { AppState } from '../services/discover_state';
 import { FetchStatus } from '../../../types';
 import { DataPublicPluginStart } from '../../../../../../data/public';
-import { SavedSearchData } from './use_saved_search';
+import { SavedSearchData } from '../services/use_saved_search';
 import { DiscoverServices } from '../../../../build_services';
 import { ReduxLikeStateContainer } from '../../../../../../kibana_utils/common';
