+++ src/plugins/discover/public/application/apps/main/utils/fetch_total_hits.test.ts
